+++ labelbox/__init__.py
@@ -1,5 +1,5 @@
 name = "labelbox"
-__version__ = "3.28.0"
+__version__ = "3.29.0"
 
 from labelbox.client import Client
 from labelbox.schema.project import Project
@@ -27,3 +27,4 @@ from labelbox.schema.iam_integration import IAMIntegration
 from labelbox.schema.resource_tag import ResourceTag
 from labelbox.schema.project_resource_tag import ProjectResourceTag
 from labelbox.schema.media_type import MediaType
+from labelbox.schema.slice import Slice, CatalogSlice